fixed bug in delete button for user of assignment and showing assignment path

+++ js/show-user-assignment.js
@@ -37,41 +37,46 @@ $(document).ready(function () {
 
   $('[data-toggle="tooltip"]').tooltip();
   // Delete row on delete button click
-  $(document).on("click", ".delete", function () {
-    const user_id = document.querySelector('.delete').id;
-    console.log("User_id", user_id)
-    $.ajax({
-        url:`https://course.simplebar.dk/api/assignment/${assign_id}/student/${user_id}`,
-        type: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${sessionStorage.getItem("token")}`,
-        },
-   //<-----this should be an object.
-        contentType:'application/json',  // <---add this
-        dataType: 'json',                // <---update this
-        success: function (result) {
-          console.log("SUCESS OF DELETE ASSIGNMENT API \n");
-          console.log(result);
-          course_details = result;
-          Swal.fire({
-            icon: "success",
-            title: "Done",
-            text: `User  ${user_id} has been deleted successfully!`,
-            footer: "Click OK to go back to dashboard",
-          }).then(() => {
-            $(this).parents("tr").remove();
-            $(".add-new").removeAttr("disabled");
-          });
-        },
-        error: function (result) {
-          console.log("FAILED DELETE ASSIGNMENT API CALL", result);
-          Swal.fire({
-            icon: "error",
-            title: "Unable to delete the user",
-            text: `${result.responseJSON.message}`
-          })
-        }
-    });
+  $(document).on("click", '.delete', function () {
+    var result = confirm(`Are you sure you want to delete user named: ${$(this).closest('tr').find('td:eq(1)').text()} ?`);
+
+    if (result) {
+      const user_id = document.querySelector('.delete').id;
+      console.log("User_id", user_id)
+      $.ajax({
+          url:`https://course.simplebar.dk/api/assignment/${assign_id}/student/${user_id}`,
+          type: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${sessionStorage.getItem("token")}`,
+          },
+     //<-----this should be an object.
+          contentType:'application/json',  // <---add this
+          dataType: 'json',                // <---update this
+          success: function (result) {
+            console.log("SUCESS OF DELETE ASSIGNMENT API \n");
+            console.log(result);
+            course_details = result;
+            Swal.fire({
+              icon: "success",
+              title: "Done",
+              text: `User  ${user_id} has been deleted successfully!`,
+              footer: "Click OK to go back to dashboard",
+            }).then(() => {
+              $(this).parents("tr").remove();
+              $(".add-new").removeAttr("disabled");
+            });
+          },
+          error: function (result) {
+            console.log("FAILED DELETE ASSIGNMENT API CALL", result);
+            Swal.fire({
+              icon: "error",
+              title: "Unable to delete the user",
+              text: `${result.responseJSON.message}`
+            })
+          }
+      });
+
+    }
 
   });
 });
